Remove Indexer class
The indexer class does not belong into the fulltext init file, but
instead of moving it into an indexer file we can just remove it and
place all needed functions into the sql indexer.

The original idea for the base indexer class was to support multiple
indexers but currently we have no plans to do so, and that leaves us
with just excessive code.