Fix downgrade for assignee migration
The downgrade migration should leave the database in the same state as
it was before the upgrade.

Not doing a proper downgrade and then having hacks in the upgrade
procedure to ensure it works from both different states is bad practice
and should be avoided. This commit will allow us to fix the normal
upgrade part of the migration.