Remove order by __similarity__ for assessments
Since changing the default order of related assessments and issues to
finished date, there is no way for a user to sort assessments by
similarity. This makes all the sorting code dead code that prevents some
performance improvements. This commit just removes the code that is
currently not accessible from our front-end.