Remove extra queries for unmapping handlers
When unmapping objects with imports we trigger a query to get the
latest relationship objects from the db, which we expect to be there
after a flush or a commit. Using the mapper helps by removing the extra
query and allows us to avoid having separate commits or flushes to sync
the data.