Add Extra ignore check for mapping columns
When there is an error in a given import row the object must be ignored.
Before that was done by having separate commits and we would never try
to import additional objects for ignored rows, but if we try to remove
those extra commits we have ta add this additional check.