Remove redundant db session flush
The db session in each cell was used for safety before. Now all checks
for proper objects should be done when parsing the data so the flush is
not needed anymore. Note that safety just means, we want to fail as soon
as possible to remove a single object in a block when there's an error
instead of having the entire CSV block fail.