Add email and names to acl json responses
This is a quick way to return needed person data to use acl without
additional queries. The proper solution would be to add "include_links"
for person to ACL but that would return too much data. Other better
option would be to add the missing data to person stubs, but that would
mean too many changes to our json builder.