Remove js file from license header exceptions.
The javascript file had it's header added and tests run before the check
for headers was merged and because we don't force all branches to be up
to date this was a conflict that was missed by automated tests.

This fix just makes that we will now have a valid header checks.

Note: the proper thing that should be done in order to avoid this clash,
would be to re-run tests on all branches right after the checks for
license headers were merged.