Add type field to revision content
Revision content is being used more and more on the front-end instead of
fully serialized objects, due to performance issues and with snapshots
because there is no other option. Because there is no type info in
current content most of the front-end code that handles revision content
is special and fetches object type data from elsewhere. With this commit
we remove the need for special handling of this data and make sure that
the front-end can use existing code for handing models.

One example for the current use case is returning issues with assessment
custom endpoint `related_objects`.