Update snapshot export tests
These updates do not change the tests or its results, but do make the
output cleaner when a test fails. This is because unicode and normal
strings match normally, but when displaying the diff between dicts, the
"u" prefix makes a lot of the lines show up as different even though
they count as the same value.