Increase priority of snapshot content data
When exporting we should prioritize the data that is already present in
the revision content and only if it's not there should we look into
cached objects.

This commits switches the order between the first two if statements and
replaces the elif statements with if statements because the result is
the same.