Update revision content for documents
We must makes sure that the display name is used for getting what is
shown for evidences and not slug. This commit is needed to fix exporting
of snapshots with evidences.

Before this commit evidences on snapshots were exported by using the
slug attribute that was queried and that original object could be
removed and thus deleted from the snapshot. By updating how we use
display names, we can avoid having special cases inside snapshot export
and still get the correct csv values.