Remove auditor role permissions
Because we couldn't remove audit contexts in scope of audit acl the
auditor role needed to be kept around until we remove them completely.
The side affect was that old auditors set before the migration would not
get properly removed after the migration (acl records got removed, but
the user roles didn't) this caused phantom auditor permissions to remain
in the app. This fix solves the problem by removing all permissions from
the old auditor role. After we remove Program and Workflow user roles we
will be abel to remove the roles completely.