Make symlink absolute
This commit just makes a symlink absolute to match the path of
node_modules symlink. In some cases docker fails to build if the folders
referenced by the symlinks do not exists and having both of our links
absolute makes it easier to mock the folder structure so that the
container builds properly.