Move fulltext join counter
The _count variable is used to enumerate multiple instances for joining
fulltext table onto a given object. Since the count must be increased
every time that fulltext table is joined, it is a lot safer to increase
the counter where it's needed and not rely that we would increase it
before every call to by_fulltext function.