Reduce scope of joined query counter
The counter for joined queries is only needed for order by joins on
fulltext table and should be encapsulated in the function that applies
order by clause. This will allow us to move the order by function into
the utils module for it to be used by other special API endpoints.