Update pagination count query
The optimization with query statement does not bring any measurable
improvements, but they do show a lot of warnings if used with any custom
query, like applying limit on an eager query with links included.

With this change we just make sure that the limit function can be safely
applied on any query in our app.