Fix assessment generation with missing roles
In assessment generation we should we should ignore any ACL entry for a
role that was deleted, because those roles can not be selected for
default assessment roles. For those we only use non editable roles which
should always be in our db. This fix ensures that assessment generation
will not fail due to unneeded data missing.