Refactor program relationship propagator
Instead of having some rules that propagate up (old comment_or_document)
we now make sure to always propagte permissions down. This is why we
have added a new Audit rule that was not there before and we have also
split up Assessment/Issue as well. Assessments template are now handled
in the audit block.