Remove invalid acl propagation checks
The tests were written with the wrong propagated roles parent
implemented and are now obsolete since the role propagation has been
fixed.

The bug was because all propagated roles for cycles, cycle tasks, and so
on had the same parent role which belongs to workflows. The correct
propagation parents should have workflow roles only on cycles. Cycle
task groups must have cycle roles as parents, cycle tasks should have
cycle task group roles as parents, and so on.

This commit ensure we still check the correct number of roles, but it
does not check for proper parents.