Fix related assessments count
It seems that sa.func.count works on polymorphic tables but does not
work properly on eager queries. While query.count() works on eager
queries and not on polymorphic tables.

This fix just ensures we get the correct counts on all pages, but is not
the proper fix for the underlying problem.