Move Issue tracker statuses map
The issue tracker statuses map should be a public constant if it's being
used in multiple places.

The reason for moving it out of the hooks file is because it causes
hooks to be imported before all models are imported and that breaks
causes problems when trying to consolidate hooks from different objects.
One example is moving workflow hooks into ggrc module.