Remove is_anonymous check from comment hook
The is_anonymous check was added here because tests use anonymous user
to do a few things, but that was a bug in the tests and not in the code.
The code should never be access in our application if an actual user is
not logged in. That is why it is better and safer to have an exception
thrown than to allow creation of invalid comments (without an admin).