Replace obsolete roles cache function
The get_custom_roles_for function works with only names and ids and
should be replaced with get_ac_roles_for that returns actual ORM objects
that can be used in the mapper. The new function also maps names to
objects instead of ids, so we can avoid redundant loops through all
comment roles to just find the admin role.