Move workflow ACL hook to ggrc module
Since we deprecated pluggable modules it should be fine to have workflow
code in the ggrc module. In this case we are doing it to join all acl
propagation into a single efficient operation. Once stop loading modules
dynamically we can move the code for the workflow hook back into
ggrc_workflows.