Remove redundant relationship creation
The relationships created between original objects and audits were done
as a hack to ensure the unified mapper showed already created snapshots
as mapped. This issue was later fixed on the front-end but the legacy
code was never removed. Now the code just presents a performance
degradation and brings no actual value to our app.

Second thing to consider is that the relationships are generated ONLY on
audit creation from the objects mapped to the programs. Any additional
snapshots that are created and mapped to the audit at later time, do not
have these relationships. And since half of the snapshots have them and
half don't without showing any bugs in our app, we can safely say that
removing this code should not cause any regressions. Any bugs that would
be caused by this must already exist since we can create the same state
already depending on how a snapshot is created.

This commit also removes the checks for those relationships and their
revisions.