Ignore all propagated roles on relevant filters
The propagated roles are only used to determine if a user has access to
a given object or not. They are not user visible roles and should not be
used in any user visible filter. The check for parent_id ensures that
only base roles are used by the query.