Fix computed attributes with relationships
It appears that we do not need to compute new attribute values on
mapping a snapshot of a control to an assessment. This is because
mapping to an assessment will move the assessment to "In Progress" state
which is ignored when computing the last assessment date. The only place
where a mapping can be done to a finished assessment is on import, but
that already creates a new revision and handles the assignment
correctly.

The original solution was:

diff --git a/src/ggrc/data_platform/computed_attributes.py
           b/src/ggrc/data_platform/computed_attributes.py
index 2ced33763a..7bb81f26a8 100644
--- a/src/ggrc/data_platform/computed_attributes.py
+++ b/src/ggrc/data_platform/computed_attributes.py
@@ -629,7 +629,9 @@ def get_revisions(revision_ids):
       models.Revision.resource_type,
       models.Revision.resource_id,
       models.Revision.source_type,
+      models.Revision.source_id,
       models.Revision.destination_type,
+      models.Revision.destination_id,
   ).filter(
       models.Revision.resource_type != "Snapshot",
       models.Revision.id.in_(revision_ids)

Which also fixes the issue but further investigation proved that the
code with the bug was dead code and not used in later parts of the
calculation. Given all this, it should be safe to remove the buggy code
without doing fixes for using the relationships to snapshots.

The extra benefit of the current solution is also slight performance
gain on generating assessments.