Fix factory usage in integration tests
Some places used wrongly imported models instead of factories for those
models. The mistake was not visible because related objects are all
pulled into a session and the objects were still created. But the end
result will be missing revisions and other parts of the model instance.
If we wish to use actual models in the future we should explicitly
import them from the models module and not from factories.