Fix imports in integration test factories
We should prevent the user to directly import the model from factories
module instead of the factory itself. This mistake has popped up in some
tests where AccessControlList model was used instead of
AccessControlListFactory. The changes in this commit make sure that no
actual models can be directly imported or used from the factories
module.