Create program audit relationships on new audit
This ensures that we have a relationship object created between a
program and an audit when creating a new audit.

The skip for acl propagation on relationship is there because audits
have a special propagation code already implemented, which will also be
removed once we have all propagated objects connected with
relationships.