Add extra logging on API exceptions
The translate message just strips out parts of the message so that it
can be displayed on the front-end, but the exception itself is not shown
anywhere. This change makes it a lot easier for developers to track down
errors cause the request to fail.

Additionally without this our logs would not log any errors, just
warnings when an API request fails due to duplicate key.