Simplify workflow ACL deletion code
The code to delete related ACLs is a hack but still needed for now. This
commit makes the delete query simpler and optimizes it for proper index
usage. This also ensures we can attach more objects to the delete queue
with a simple set union instead of dealing with individual items in a
default dictionary.