Fix license header checks
This commit ensures that the exact string containing license header
information is not present in the script that checks for it. The issue
is a false negative if we were to remove the licence header comment from
the top of the script that checks for it. Before this change the script
would match the string in the check instead of the text on top of the
file that would be missing.