Remove check_flake8_diff script
The diff checks were used only for comparing flake 8 on files that were
modified. Now that we have removed all current flake 8 issues from our
codebase, we can just run flake 8 on the entire thing and use that
result. This will make it easier to check files locally as well.