Add SQL propagation for ACL entries
This file enables us to propagate entire layers of acl propagation tree
with a single sql statement.

The maximum amount of queries triggered by this propagation is defined
by the depth of the propagation tree, with a safety cut-off at depth 50.

Example of propagation for

Auditors propagation rule:

Audit
  - relationship
      - assessment
         - relationship
            - issue
            - document

In the example the code below would receive a new auditor acl id and
propagate the correct permissions to relationships, assessments, issues,
and documents in 4 sql queries.