Fix propagation conflicts
When we create a new acl on an object and a relationship on the same
object in the same commit, we get a conflict. This happens because
relationships will try to propagate all roles of connected object, and
normal acl propagation will try to re-do the propagation only for the
new acl entries. With having an exclude condition we ensure that all
items will be propagated only once and we can avoid having insert into
ignore statement.