Fix issue mapping tests
We should never use propagated roles in our app, the proper thing is to
use the original role and let propagation code do the right thing. This
is because propagation can change and as long as the permissions are
correct this test should not fail.

Besides that we have ensured that program-audit and audit-snapshot
relationships are always created since the new propagation code relies
on those. That is why a proper setup must also include those
relationships.

The test fix includes:

- Remove usage of program internal roles.
- Add mapping between snapshot and audit.
- Add mapping between program and audit.