Change Auditors propagation rules
The specification on audit propagation rules says that auditors should
be able to map things to issues and that they should not have edit
access to any other audit scope objects which is a contradiction. Since
the rules can't be implemented in our current codebase I have opted to
set the same propagation rules that were implemented before the ACL
propagation update.