Remove internal role asserts from tests
Our internal roles have changed and it is not valid to have them in
propagation tests. The only thing that should be tested is what
permissions a person get's on an object.

Because we have a separate test suite for permissions that should cover
these cases completely, it should be safe to just remove these asserts.