Fix assessment model tests
This fix ensures we check for propagated role by their parent and not
the current role names.

The more proper check would be to step through ACL parents and then
check the top most parent type, but that is left out since we have
different ACL test cases for checking correct role propagation.