Change document propagation rules
Currently in our app we never delete documents. A delete action is done
by unmapping a document from an object. This was also a decision when
doing the original propagation rules. But since we have a test that
ensures some users can also delete the actual document, we have decided
to change propagation rules instead of the tests.

This does not have a big impact on the app, as users will eventually get
delete permissions on documents and the entire document design will be
modified.