Fix global propagation function
The splitting of propagation did not work before, because we are
updating and extending the table that we have split into chunks. This
means that the last chunk will also try to propagate some of already
propagated roles and it will result in unique constraint error.

Loading all ACL ids and splitting those integers into chunks seems safe
enough for now.