Add ACL downgrade option with info
This change allows us to have a semi-proper downgrade with info on all
introduced changes by the given downgrade.

Note: this commits adds two warnings on the original upgrade procedure
that can be ignored. Sqlalchemy warns about unknown table due to "delete
if not exists" clauses. But the warning is invalid and the migration
still works and subsequent migrations also work without the warnings
once the table has been created.