Add parent_id_nn to access control list model
The parent_id_nn or parent id not nullable column on access control list
is just a duplicate of parent_id column but with a default value of 0
and without fk and index on it.

This column will be used for joins, to prevent mysql from using the
wrong index (joining on object_id, object_type, parent_id, only used the
parent_id index and was too slow). This could also be achieved with
`force index on join` but we can't add that to sqlalchemy relationship.
The added benefit of this column is then to also help preventing
duplicates on the unique constraint on that table.