Remove loading of obsolete permissions
Backlog workflows have never been used and full access permissions on
those are not needed anymore.

Since we don't have the ability to create a backlog workflow anymore, we
can be sure that his will not cause rbac issue in the future.

The following statement was run on all deployed databases with the same
result:

> select count(*) from workflows where kind is not null;
+----------+
| count(*) |
+----------+
|        0 |
+----------+