Remove unicode warnings from attributes deletion
Since the object type and id are set automatically by ORM mapper when
deleting an assessment we can't force it to be unicode and then
sqlalchemy shows a warning that we're comparing ascii to unicode.

The original reason for using unicode was to be as close to data
platform design as possible, but the key difference here is that our
models are actual python classes that can't be unicode. This makes it
safe to change the filed from unicode to normal string.