Add special acl handler for cycle task comments
Originally to avoid handling all special cases, any new object that was
added to a workflow scope, triggered the entire workflow propagation.
This is good enough for cycle generation and other such operations
because we avoid duplicating work, or specifying all possible special
cases for child propagation, when multiple objects are created in a
single transaction.

Cycle tasks on the other hand need to be faster and they are easy to
handle because they do not trigger any new propagation - they are the
leafs in the propagation tree.

This commit pulls out all cycle task ids on which new comments have been
added. Then it triggers only cycle task entry propagation on those cycle
tasks. The work still duplicates any propagation of the previous
comments but the impact of that is negligible.