Fix full acl propagation for workflows
This is a fix caused by a regression in the commit af4173508e4674. The
offending line 42 in workflow hook makes sure we have all three wf
specific variables set before running the propagation, but in full
propagation we forgot to set the newly added wf comments variable.

The change of wf acl ids from list to set is there just for consistency
since the code should always expect to get a set of ids.