Split acr propagation deletion into chunks
Having a single big delete statement that takes a long time is hard to
track in the logs if something went wrong and it has a higher risk of
triggering lock_wait_timeout error.

Moving the delete statement to right before we propagate we also reduce
the window in which user interaction can create conflicts on
propagation. Conflicts can still appear if users create a new mapping on
the slice of objects that is being re-propagated.