Reduce test execution time for api requests
Originally we had setupclass method that created all data needed for
executing all tests. Then we avoided calling super setup on every test.
This is now no longer possible since setup method initializes memcache
that is used when importing a file. This commit just makes sure our test
execution is almost as fast as it was before but at the expense of only
showing the first assertion error if there are multiple.

Better fixes for this test will be available once we fix the setup
method to allow persistent data over multiple tests.