Remove invalid assertion from tests
The test for concurrency was meant to check that acl propagation does
not cause any problems, but we forgot about the actual etags that do not
match on different concurrent requests. So it is valid that some of the
requests might fail, but we still check that the end result of the
requests is correct.

The error hidden in this test started to appear after we added a strict
unique constraint on acl table - meaning no columns in the constraint
can be nullable. The error is triggered in the main commit loop before
the propagation even starts.

We must still fix concurrency issues here, but acl propagation is
not the cause of the errors in this test.