Remove context_query_filter from query API
The main change in this commit is removal of context_query_filter which
can only return sa.true() in the current state. But the code is also
refactored to avoid returning a filter if it's always true since query
without a filter will return the same results as a query with always
true filter.