Remove context_query_filter from joined queries
This commit is just a refactor of the code to remove the "or true"
statement caused by context_query_filter on non None contexts list.
The point of this commit is to refactor the code with the exact same
results as before, but since this code does not seem to be tested I have
not tried to fix its weird behavior.