Clean up variable names in app.py
The invalid name for logger was actually used in the entire file scope
and since we have added logger into pylintrc exceptions we should not
disable the name check in the file. This commit just adds the valid
exception for the name "app" which is only for the given line, and fixes
the invalid name "e".