Move indexer benchmark statement
We should avoid logging the benchmark if there is nothing to be indexed.
The actual use for the statement before the return statement would be to
check that this function is not called too many times. But we have other
fail-safes for too many commits.

This commit just prevents the spamming of logs on total acl propagation
recalculation.