Remove obsolete backlog workflow endpoint
The backlog workflows were removed in scope of pr #7773 for reducing
loading for permissions. The check at the time proved that no backlog
workflows exist in any prod or dev environment so permission checks for
those were removed. Having an endpoint that creates such a workflow is
now considered a regression so it has to be removed.