Separate SELECT and INSERT on ACL propagation
INSERT INTO a SELECT FROM b puts read locks on b, which leads to
deadlocks when the same statement is executed 3 times in a row by each
parallel job with slight variations.

SELECT, COMMIT and then INSERT effectively reduces the transaction
isolation, because SELECTed data that stays in the app may go obsolete
right after the new transaction is opened. For the case of ACL
propagation this is good enough, because any change in the data isn't
going to affect the result of the INSERT. The only possible issue is a
race condition: req1/tx1 SELECTs data based on an ACL entry, parallel
req2/tx1 drops the ACL entry, and req1/tx2 tries to INSERT a row that
will violate FK constraint.