Remove acl propagation utils
The main function for inserting propagated entries was in the utils file
because it was used for the main propagation and workflow propagation.
Since we have removed workflow propagation there is no need for a
utilities module with a single function that is used in only one file.

This change only moves inserter function and makes it private in the
file that it belongs to. There are no code changes in the function
itself.