Update initial migration
There should not be any objects without revisions on an empty database.
The entry deleted here was due to a bug on how the old migration worked
on an empty db, when it created an invalid entry.

Second change is updating the title of the migration to display the
correct version from which the squash was created.