Fix test assertions
The change in test assertions make sure that in case of a failed test
we don't get raise an error but show a failed test with the proper
message on what fails. The new assertions also prevent any unwanted
warning messages from showing up which could have been missed before.