Wrap test client calls into app context
Not having individual context for each call means that flask.g variables
persist over multiple tests and multiple requests. Because we need all
test runs to be fully independent, and because requests must behave as
if the app was running we should ensure app context and cleanup for all
requests.