Update test query limit
The event and revision query limit should be the same. The reason why
this worked before was due to invalid clearing of caches in test client
calls. Since flask.g cache remained from the first check the second
request showed as one fewer queries, but in a normal app the counts
would have been the same so previous tests were invalid.

Since we wrapped all client requests into an app context that clears
flask.g this test began failing and this commit now sets the correct
number to check.