Fix access control base function
This is actually invalid comparison, but the document test cases send
document admin role to create a user role for the parent object. This
only worked because the role names matched for document and parent Admin
but the correct thing would be to send the correct role when assigning a
person to the document parent object.