Fix imports with error messages
Note that this fix is only for our test framework since frontend does
not allow actual import if the analysis showed any errors. Our tests on
the other hand use dry_run=False on all imports.

The bug fixed here is due to new ACL handling. On any new object
creation the ACL entries for all roles on that object get created. This
then automatically pulls the new object into the current session and in
imports that should not have been done. One helpful thing is that now we
handle every import row in its own transaciton. So instead of removing
the object from the session we can just rollback the entire session.