Remove invalid test
This test even, though it's under snapshot indexing and it creates
snapshot, does not check anything regarding snapshots. The assertions
are based on index of original object. Since it is now impossible to
create ACL entries with the wrong role (since those are created
automatically) the test to check indexing with the wrong role has become
invalid.