Force reloading of permissions for automappings
We optimized our permissions loader to only load permission for objects
referenced in the request. Since auto mappings also affect objects that
are not listed in the request, we must use old style loading of
permissions. This is is done as a hack in auto mappings because this is
the only code that should be affected.