Move personal context creation
The hook to create personal context is only needed once when the person
is created. There is no reason to call and check for that context on
every flush when a person is present in a session. The hook also causes
performance degradation if the person context is not loaded and it
lazily loads them one by one.

Note: this code will be removed soon anyway as we are in the process of
removing context model from our app.