Fix proposal test api calls
Using the api.put with the proper dict that contains the object name
will prevent the full payload to be sent. By specifying only data that
needs to be modified by the put request we ensure that the proposal
payload will contain all needed field such as id any type, that are now
mandatory for a proper acl check.