Remove redundant acl role adding optimization
Since the hook for adding missing acl roles is used when creating a new
ACR and when running propagated_acl it was made to support a filter only
for propagate_acl command. This was a performance optimization that
caused problems in the past. Now since the limit has be modified to
support the filter we should not be using the hook without the filter
for handled acl entries.

This commit resolves a problem with adding a new ACR by removing a
redundant performance improvement.