Fix CAD setter for assessment templates
This fix makes sure that `is_lca` function on CADs works properly when
creating a person local custom attribute on assessment template. Without
this change the check for local CA would fail and it would prevent
creating an assessment template with a person CA.

More info: when creating LCA on AT, the definition id is not available
because both objects are created in the same session. Id would only be
available after flush. So to ensure the rest of the code works correctly
before flush event, we have to assign the definition in the same way as
the rest of our orm mapped objects.