Add cleanup of dead records to reindex
Since we modified reindex to just update records that we have in our
database, we lost the chance to clean up any records that should not
exist. These records are a result of different bugs where a deletion of
an object fails after deletion but before the fulltext update. So here
we just make sure that after running reindex no dead objects will remain
in the fulltext index.