Fix setting edit mode for custom roles for Assessment after updating status
Custom roles are fully repainted after object refreshing.
It leads to an issue on assessment:
when user clicks on edit custom roles button and confirmation modal appears,
and user confirms changing status
then Assessment status changes to In Progress (Assessment refresh happens),
custom roles are repainted
and moving to edit mode is not happening because of repainting

Updating roles in groups instead of recalculating groups fixes this issue.