Fix ACL in test_assessment_notifications
Before 7448f25 if assessment source hasn't any snapshot in it or no
"_generated" keyword was provided, `relate_assignees` was not called.
And `relate_assignees` assignes current user as assessment creator, so
no creator was assigned, but since FE still sends one (and it is not
possible to create assessment without creator from FE side), creator
was added anyway.

In test_access_control_list creator nor snapshot nor "_generated" where
sent during assessment creation so no creator was assigned for it.

After 7448f25 `relate_assignees` is called even if no snapshot or
"_generated" keyword was provided in assessment source. Which leads to
current user being assigned as creator. And it is how it should be.

But now absence of creator in ACL being sent in PUT in test leads to it
removal and test failure. This is why creator was added there.